EAMxx: split atm procs test into smaller unit tests
