--- components/eamxx/src/share/atm_process/tests/atm_process_tests_named_procs.yaml
@@ -1,21 +1,0 @@
-%YAML 1.1
----
-atm_procs_list: [MyFoo,BarBaz]
-schedule_type: sequential
-type: group
-
-MyFoo:
-  type: Foo
-  grid_name: point_grid
-BarBaz:
-  type: group
-  schedule_type: sequential
-  atm_procs_list: [MyBar,MyBaz]
-
-  MyBar:
-    type: Bar
-    grid_name: point_grid
-  MyBaz:
-    type: Baz
-    grid_name: point_grid
-...
